requirements.txt
# local package
-e .

# external requirements
click==8.1.3
Sphinx
coverage
awscli
flake8
numpy==1.24.1
python-dotenv==0.21.0
setuptools==65.5.0
torch==1.13.1
torchvision==0.14.1
numpy==1.24.1
pytest==7.2.0
